feat: add unsetContent with UnsetContentParams to all storage modules allowing content to be kept

+++ spring-content-azure-storage/src/main/java/internal/org/springframework/content/azure/store/DefaultAzureStorageImpl.java
@@ -22,10 +22,7 @@ import org.springframework.content.commons.mappingcontext.ContentProperty;
 import org.springframework.content.commons.mappingcontext.MappingContext;
 import org.springframework.content.commons.property.PropertyPath;
 import org.springframework.content.commons.repository.SetContentParams;
-import org.springframework.content.commons.store.AssociativeStore;
-import org.springframework.content.commons.store.ContentStore;
-import org.springframework.content.commons.store.GetResourceParams;
-import org.springframework.content.commons.store.StoreAccessException;
+import org.springframework.content.commons.store.*;
 import org.springframework.content.commons.utils.BeanUtils;
 import org.springframework.content.commons.utils.Condition;
 import org.springframework.content.commons.utils.PlacementService;
@@ -460,7 +457,22 @@ public class DefaultAzureStorageImpl<S, SID extends Serializable>
     @Transactional
     @Override
     public S unsetContent(S entity, PropertyPath propertyPath) {
+        return this.unsetContent(entity, propertyPath, UnsetContentParams.builder().build());
+    }
 
+    @Transactional
+    @Override
+    public S unsetContent(S entity, PropertyPath propertyPath, org.springframework.content.commons.repository.UnsetContentParams params) {
+        int ordinal = params.getDisposition().ordinal();
+        UnsetContentParams params1 = UnsetContentParams.builder()
+                .disposition(UnsetContentParams.Disposition.values()[ordinal])
+                .build();
+        return this.unsetContent(entity, propertyPath, params1);
+    }
+
+    @Transactional
+    @Override
+    public S unsetContent(S entity, PropertyPath propertyPath, UnsetContentParams params) {
         ContentProperty property = this.mappingContext.getContentProperty(entity.getClass(), propertyPath.getName());
         if (property == null) {
             throw new StoreAccessException(String.format("Content property %s does not exist", propertyPath.getName()));
@@ -470,7 +482,7 @@ public class DefaultAzureStorageImpl<S, SID extends Serializable>
             return entity;
 
         Resource resource = this.getResource(entity, propertyPath);
-        if (resource != null && resource.exists() && resource instanceof DeletableResource) {
+        if (resource != null && resource.exists() && resource instanceof DeletableResource && params.getDisposition().equals(UnsetContentParams.Disposition.Remove)) {
 
             try {
                 ((DeletableResource)resource).delete();
@@ -488,8 +500,8 @@ public class DefaultAzureStorageImpl<S, SID extends Serializable>
                     if ("jakarta.persistence.Id".equals(
                             annotation.annotationType().getCanonicalName())
                             || "org.springframework.data.annotation.Id"
-                                    .equals(annotation.annotationType()
-                                            .getCanonicalName())) {
+                            .equals(annotation.annotationType()
+                                    .getCanonicalName())) {
                         return false;
                     }
                 }
@@ -502,7 +514,7 @@ public class DefaultAzureStorageImpl<S, SID extends Serializable>
         return entity;
     }
 
-	private String absolutify(String bucket, String location) {
+    private String absolutify(String bucket, String location) {
 		String locationToUse = null;
 		Assert.state(location.startsWith("azure-blob://") == false);
 		if (location.startsWith("/")) {
