+++ spring-content-gcs/src/main/java/internal/org/springframework/content/gcs/store/DefaultGCPStorageImpl.java
@@ -21,6 +21,7 @@ import org.springframework.content.commons.mappingcontext.ContentProperty;
 import org.springframework.content.commons.mappingcontext.MappingContext;
 import org.springframework.content.commons.property.PropertyPath;
 import org.springframework.content.commons.repository.SetContentParams;
+import org.springframework.content.commons.repository.UnsetContentParams;
 import org.springframework.content.commons.store.AssociativeStore;
 import org.springframework.content.commons.store.ContentStore;
 import org.springframework.content.commons.store.GetResourceParams;
@@ -456,46 +457,59 @@ public class DefaultGCPStorageImpl<S, SID extends Serializable>
     @Transactional
     @Override
     public S unsetContent(S entity, PropertyPath propertyPath) {
-
-        ContentProperty property = this.mappingContext.getContentProperty(entity.getClass(), propertyPath.getName());
-        if (property == null) {
-            throw new StoreAccessException(String.format("Content property %s does not exist", propertyPath.getName()));
-        }
-
-        if (entity == null)
-            return entity;
-
-        Resource resource = this.getResource(entity, propertyPath);
-        if (resource != null && resource.exists() && resource instanceof DeletableResource) {
-
-            try {
-                ((DeletableResource)resource).delete();
-            } catch (Exception e) {
-                logger.error(format("Unexpected error unsetting content for entity %s", entity));
-                throw new StoreAccessException(format("Unsetting content for entity %s", entity), e);
-            }
-        }
-
-        // reset content fields
-        property.setContentId(entity, null, new org.springframework.content.commons.mappingcontext.Condition() {
-            @Override
-            public boolean matches(TypeDescriptor descriptor) {
-                for (Annotation annotation : descriptor.getAnnotations()) {
-                    if ("jakarta.persistence.Id".equals(
-                            annotation.annotationType().getCanonicalName())
-                            || "org.springframework.data.annotation.Id"
-                                    .equals(annotation.annotationType()
-                                            .getCanonicalName())) {
-                        return false;
-                    }
-                }
-                return true;
-            }
-        });
-        property.setContentLength(entity, 0);
-
-        return entity;
+		return this.unsetContent(entity, propertyPath, org.springframework.content.commons.store.UnsetContentParams.builder().build());
     }
+
+	@Override
+	public S unsetContent(S entity, PropertyPath propertyPath, UnsetContentParams params) {
+		int ordinal = params.getDisposition().ordinal();
+		org.springframework.content.commons.store.UnsetContentParams params1 = org.springframework.content.commons.store.UnsetContentParams.builder()
+				.disposition(org.springframework.content.commons.store.UnsetContentParams.Disposition.values()[ordinal])
+				.build();
+		return this.unsetContent(entity, propertyPath, params1);
+	}
+
+	@Override
+	public S unsetContent(S entity, PropertyPath propertyPath, org.springframework.content.commons.store.UnsetContentParams params) {
+		ContentProperty property = this.mappingContext.getContentProperty(entity.getClass(), propertyPath.getName());
+		if (property == null) {
+			throw new StoreAccessException(String.format("Content property %s does not exist", propertyPath.getName()));
+		}
+
+		if (entity == null)
+			return entity;
+
+		Resource resource = this.getResource(entity, propertyPath);
+		if (resource != null && resource.exists() && resource instanceof DeletableResource && params.getDisposition().equals(org.springframework.content.commons.store.UnsetContentParams.Disposition.Remove)) {
+
+			try {
+				((DeletableResource)resource).delete();
+			} catch (Exception e) {
+				logger.error(format("Unexpected error unsetting content for entity %s", entity));
+				throw new StoreAccessException(format("Unsetting content for entity %s", entity), e);
+			}
+		}
+
+		// reset content fields
+		property.setContentId(entity, null, new org.springframework.content.commons.mappingcontext.Condition() {
+			@Override
+			public boolean matches(TypeDescriptor descriptor) {
+				for (Annotation annotation : descriptor.getAnnotations()) {
+					if ("jakarta.persistence.Id".equals(
+							annotation.annotationType().getCanonicalName())
+							|| "org.springframework.data.annotation.Id"
+							.equals(annotation.annotationType()
+									.getCanonicalName())) {
+						return false;
+					}
+				}
+				return true;
+			}
+		});
+		property.setContentLength(entity, 0);
+
+		return entity;
+	}
 
 	private String absolutify(String bucket, String location) {
 		String locationToUse = null;
